# shared/openapi.taskbridge.v1.json
{
  "components": {
    "schemas": {
      "ChecklistItem": {
        "properties": {
          "done": {
            "default": false,
            "title": "Done",
            "type": "boolean"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "ChecklistItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username_or_email": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username Or Email",
            "type": "string"
          }
        },
        "required": [
          "username_or_email",
          "password",
          "device_id"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "ProductEventCreate": {
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "device_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "name": {
            "pattern": "^[a-z][a-z0-9_]{2,63}$",
            "title": "Name",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "route": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "source": {
            "enum": [
              "web",
              "desktop",
              "android"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source"
        ],
        "title": "ProductEventCreate",
        "type": "object"
      },
      "RefreshTokenRequest": {
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "refresh_token": {
            "minLength": 16,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "type": "object"
      },
      "RevokeOtherSessionsRequest": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "RevokeOtherSessionsRequest",
        "type": "object"
      },
      "SyncChange": {
        "properties": {
          "action": {
            "enum": [
              "create",
              "update",
              "delete",
              "complete",
              "restore"
            ],
            "title": "Action",
            "type": "string"
          },
          "checklist": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChecklistItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "is_template": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Template"
          },
          "list_type": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Type"
          },
          "local_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Local Id",
            "type": "string"
          },
          "local_updated_at": {
            "format": "date-time",
            "title": "Local Updated At",
            "type": "string"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "planned_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Date"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "project": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "remind_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remind Time"
          },
          "repeat_rule": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat Rule"
          },
          "server_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "sort_order": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tag": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "template_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Name"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "local_id",
          "action",
          "version",
          "local_updated_at"
        ],
        "title": "SyncChange",
        "type": "object"
      },
      "SyncPushRequest": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SyncChange"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "changes"
        ],
        "title": "SyncPushRequest",
        "type": "object"
      },
...
      "UserCreate": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "email": {
            "maxLength": 255,
            "minLength": 5,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "device_id"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebSocketTicketRequest": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "WebSocketTicketRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "TaskBridge backend API for Android and Windows desktop clients.",
    "title": "TaskBridge API",
    "version": "0.1.6"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analytics/events": {
      "post": {
        "operationId": "create_product_event_api_v1_analytics_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Product Event",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/summary": {
      "get": {
        "operationId": "read_product_event_summary_api_v1_analytics_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 366,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Product Event Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "read_current_user_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Current User",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "read_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/revoke-other-devices": {
      "post": {
        "operationId": "revoke_other_sessions_api_v1_auth_sessions_revoke_other_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeOtherSessionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Revoke Other Sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Revoke Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/ws-ticket": {
      "post": {
        "operationId": "create_ws_ticket_api_v1_auth_ws_ticket_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebSocketTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Ws Ticket",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/sync/pull": {
      "get": {
        "operationId": "pull_api_v1_sync_pull_get",
        "parameters": [
          {
            "in": "query",
            "name": "last_sync_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Last Sync Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor_updated_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor Updated At"
            }
          },
          {
            "in": "query",
            "name": "cursor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Pull",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "operationId": "push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Push",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}